DS-1241 Statistics Implementation in Elastic Search
Copied in Implementation of Elastic Search, and wired it into DSpace.
I changed the ES Node to be a nodeClient, and also one that runs locally
within tomcat's JVM. For OSU, we had used either transport-client, or
node-client, and both of which required that the server also be running a
seperate elastic search instance. By bundling this into DSpace, it should
make this a bit more turn-key to get started using Elastic Search.

By default, Elastic Search is ON with this commit.